Updated references to GitHub for contribution and README.md
References to modifying the website via GitHub ("create a pull request",
etc.) have been removed as we now host the repository on GoogleSource.
As you cannot directly modify files on GoogleSource, the "Edit this page"
feature links to the README.md file (
../public/gem5-website/+/refs/heads/master/README.md) which has been
expanded to include contribution guidelines (similar, but not identical,
to the general guidelines for contributing to the gem5 source).

Change-Id: Iffdc5e570ac9062c6e78fc6d54fc1597a288f5bc