website: Fix broken links in about page
The about page was migrated straight from the old wiki with wikilinks
formatted with the old HTML linking. This resulted in many dead links
that were incompatible with the new website. This patch changes the
links either to the correct pages on the new documentation, directly to
the old wiki, or newly-migrated pages.

Change-Id: I35f091322c7596c16878af2d99fb6cbe5a293ac5