website: Updated building docs for Python 3 support
The building doc have been updated to state that gem5 is buildable and
runable in both Python 2 and Python 3. This includes giving seperate APT
commands for when obtaining dependencies in Ubuntu 18.04 and Ubuntu
20.04.

The Ubuntu 18.04 setup is assumed to use Python 2, and 20.04 with
Python 3. This reflects our Docker and testing setup.

Change-Id: I963ad4f1a3a70e452598dccc319401c28fb64d8d
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-website/+/28868
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: Jason Lowe-Power <power.jg@gmail.com>