website: Change URL and asset path of gem5-20 benchmark status
As we are moving towards gem5-20.1, the gem5-20.1 benchmark status
page should be the landing page of the benchmark status page. The
gem5-20 benchmark status should be archived.

This commit changes the URL of gem5-20 benchmark status page so
that it is no longer associated with the default benchmark status
page. The assets used in the page are also moved to a separated
folder. This allows separations of assets among different versions
of gem5 benchmark status pages.

This commit also add a table of content to the benchmark status
page.

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>
Change-Id: I8ad3117b64b34980d9598712d51885e0469f27d5
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-website/+/35315
Reviewed-by: Bobby R. Bruce <bbruce@ucdavis.edu>
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: Bobby R. Bruce <bbruce@ucdavis.edu>