website: Provide multiple init alternatives in ubuntu image
Unfortunately it is not possible to specify a single m5 powered init
script working under every circumstance. For example when using a KVM
CPU, the default coprocessor implementation of m5 ops won't be
recognised by the host; same applies for the semihosting one; the memory
mapped version on the other hand is not compatible with a Fast Model CPU
which instead requires the semihosting interface.

To overcome these limitations, we provide a single disk image with
multiple init scripts. The user would have to point to the correct init
script depending on the simulation environment.

/init.gem5 -> If using a "standard" cpu
/init.addr.gem5 -> If using a kvm cpu
/init.semi.gem5 -> If using a fast models cpu

Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>
Change-Id: Iab58e01eea4f465d670f6dcebd0c46f4fcf55247
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-website/+/44005
Reviewed-by: Richard Cooper <richard.cooper@arm.com>
Reviewed-by: Bobby R. Bruce <bbruce@ucdavis.edu>
Maintainer: Bobby R. Bruce <bbruce@ucdavis.edu>
Tested-by: Bobby R. Bruce <bbruce@ucdavis.edu>