website: update GCN3 webpage
This commit makes two major updates to the GCN3 webpage:

1.  Removes the majority of the examples and instead points to the
examples in gem5-resources, which are both more up-to-date and
more numerous.  This eliminates the need to update documentation
in two places, and should reduce confusion moving forward.

2.  Updates the GCN3 information to reflect that ROCm 4.0 is now
supported instead of ROCm 1.6, as well as removing the ROCm libraries
that are no longer needed with ROCm 4.0.

Change-Id: I5421f3c1982c01d3b346cc91ec404e186345e4d6
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-website/+/50907
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Reviewed-by: Matthew Poremba <matthew.poremba@amd.com>
Maintainer: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: Jason Lowe-Power <power.jg@gmail.com>