website: Re-add gem5 101 content
The gem5 101 page referenced a wiki which has been taken down. The
content has been re-added by translating a text-dump of the wiki to
Markdown.

Important Note: While efforts were made to be faithful to the original
tutorials, some dead-links persist in Homeworks 3, 4, 5, and 6. It is
also possible the tutorials are uncompatible with the latest versions of
gem5. This should be treated more-so as an archive than references for
learning gem5.

Change-Id: I42ca5bf86d6c2b82e51ad93224740f8398abd434
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5-website/+/55686
Reviewed-by: Jason Lowe-Power <power.jg@gmail.com>
Tested-by: Jason Lowe-Power <power.jg@gmail.com>
Maintainer: Bobby Bruce <bbruce@ucdavis.edu>