website: Update gonvernance to replace committer with maintainer
When we used reviewboard (many, many years ago) the term for the person with the ability to modify the mainline code was "committer." Now that we are using github, we're using the term "maintainer." This commit makes that change to the governance file. Other minor reviewboard->github changes are also included.